Reactor promoted content to match new app architecture
