+++ moj-node/server/services/hubPromotedContent.js
@@ -1,20 +1,12 @@
-const { HubContentClient, HubFeaturedContentResponse } = require('../clients/hubContent');
-
-module.exports = function createHubPromotedContentService() {
-
-  async function hubContentPromotedFor(httpClient, opts = { query: {}}) {
-    const response = await httpClient.get("/promoted", opts.query);
-    if(response.message) return null
-    return HubFeaturedContentResponse.parse(response);
+module.exports = function createHubPromotedContentService(repository) {
+  async function hubPromotedContent() {
+    try {
+      return repository.hubPromotedContent();
+    } catch (ex) {
+      return null;
+    }
   }
-  
-  async function hubPromotedContent(httpClient) {
-    const client = new HubContentClient(httpClient);
-    return hubContentPromotedFor(client) 
-  }
-
   return {
-    hubContentPromotedFor,
-    hubPromotedContent
-  }
-}
+    hubPromotedContent,
+  };
+};
